Move stress tests into integration folder
